test_fb_insta_api/requirement.txt
facebook-sdk==3.1.0
nltk==3.4.5
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.2
pymongo==3.11.3
python-facebook-api==0.9.0
regex==2020.10.15
requests==2.25.1
rsa==4.6
scikit-learn==0.23.2
seaborn==0.11.1
sklearn==0.0
Unidecode==1.2.0
urllib3==1.26.3
